api/requirements.txt
Django==3.2.9
Pillow==9.0.0
PyYAML==6.0
django-filter==21.1
djangorestframework==3.12.4
psycopg2-binary==2.9.2
uritemplate==4.1.1
